Handle plain scalars
Multi-line scalars are not fully handled in the lexer. They are treated as separate lines, and we'll have to join them in the parser, or in a filtering/merging lexer. We'll also have to handle continuation lines that begin with characters that first lines aren't allowed to. E.g. c-indicator